SB13 - support chain IDs larger than type(uint32).max

+++ src/base/ChainSettlers.sol
@@ -9,12 +9,12 @@ contract ChainSettlers is Ownable2Step {
     /// @notice Error thrown when the chain settler params length mismatch
     error ChainSettlersParamsLengthMismatch();
     /// @notice Error thrown when the chain settler is not supported
-    error ChainSettlerNotSupported(uint32 chainId, address settler);
+    error ChainSettlerNotSupported(uint256 chainId, address settler);
 
     /// @notice Event emitted when a chain settler is updated
-    event ChainSettlerUpdated(uint32 indexed chainId, address indexed settler, bool value);
+    event ChainSettlerUpdated(uint256 indexed chainId, address indexed settler, bool value);
 
-    mapping(uint32 => mapping(address => bool)) public chainSettlers;
+    mapping(uint256 => mapping(address => bool)) public chainSettlers;
 
     /// @notice Constructor for the ChainSettlers contract
     /// @param initialOwner The initial owner of the contract
@@ -25,7 +25,7 @@ contract ChainSettlers is Ownable2Step {
     /// @param settlers The settlers
     /// @param values The values
     /// @dev Only the owner can call this function
-    function setChainSettlers(uint32[] calldata chainIds, address[] calldata settlers, bool[] calldata values)
+    function setChainSettlers(uint256[] calldata chainIds, address[] calldata settlers, bool[] calldata values)
         external
         onlyOwner
     {
