+++ src/interfaces/IMigrator.sol
@@ -35,7 +35,7 @@ interface IMigrator {
 
     /// @notice Struct for migration parameters
     struct MigrationParams {
-        uint32 chainId;
+        uint256 chainId;
         address settler;
         TokenRoute[] tokenRoutes;
         bytes settlementParams;
